+++ Packages/cc.sugi.imageprocesstool/Runtime/Data/ImageProcessPortDirection.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace SugiCho.ImageProcessTool
+{
+    [Serializable]
+    public enum ImageProcessPortDirection
+    {
+        Input = 0,
+        Output = 1
+    }
+}