+++ Packages/cc.sugi.imageprocesstool/Runtime/Data/ImageProcessPortType.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace SugiCho.ImageProcessTool
+{
+    [Serializable]
+    public enum ImageProcessPortType
+    {
+        Texture = 0,
+        Float = 1,
+        Vector4 = 2,
+        Color = 3
+    }
+}